rename mulh to mul
