+++ ceno_zkvm/src/instructions/riscv.rs
@@ -17,7 +17,7 @@ pub mod ecall;
 pub mod jump;
 pub mod logic;
 pub mod logic_imm;
-pub mod mulh;
+pub mod mul;
 pub mod shift;
 pub mod shift_imm;
 pub mod slt;
