+++ ceno_zkvm/src/instructions/riscv/rv32im.rs
@@ -10,7 +10,7 @@ use crate::{
             divu::DivUInstruction,
             logic::{AndInstruction, OrInstruction, XorInstruction},
             logic_imm::{AndiInstruction, OriInstruction, XoriInstruction},
-            mulh::MulhuInstruction,
+            mul::MulhuInstruction,
             shift::{SllInstruction, SrlInstruction},
             shift_imm::{SlliInstruction, SraiInstruction, SrliInstruction},
             slti::SltiInstruction,
@@ -32,7 +32,7 @@ use divu::{DivDummy, RemDummy, RemuDummy};
 use ecall::EcallDummy;
 use ff_ext::ExtensionField;
 use itertools::Itertools;
-use mulh::{MulInstruction, MulhInstruction, MulhsuInstruction};
+use mul::{MulInstruction, MulhInstruction, MulhsuInstruction};
 use shift::SraInstruction;
 use slt::SltInstruction;
 use slti::SltiuInstruction;
